--- frontend/src/components/Block.jsx
@@ -1,12 +1,0 @@
-import React from "react"
-import './../styles/Block.css'
-
-function Block() {
-    return (
-        <div className="Block">
-
-        </div>
-    )
-}
-
-export default Block
