+++ frontend/src/components/basicUser/CompleteGameBlock.jsx
@@ -1,0 +1,14 @@
+import React from "react"
+import '../../styles/basicUser/CompleteGameBlock.css'
+
+function CompleteGameBlock() {
+    return (
+        <div className="CompleteGame_block">
+            <div className="CG_title_block">
+                Пройти игру
+            </div>
+        </div>
+    )
+}
+
+export default CompleteGameBlock
