+++ frontend/src/components/basicUser/FillingFormBlock.jsx
@@ -1,0 +1,14 @@
+import React from "react"
+import '../../styles/basicUser/FillingFormBlock.css'
+
+function FillingFormBlock() {
+    return (
+        <div className="FillingForm_block">
+            <div className="FF_title_block">
+                Заполни форму <br />и прикрепи видеовизитку
+            </div>
+        </div>
+    )
+}
+
+export default FillingFormBlock
