+++ frontend/src/components/clientAdmin/Block.jsx
@@ -1,0 +1,12 @@
+import React from "react"
+import '../../styles/clientAdmin/Block.css'
+
+function Block() {
+    return (
+        <div className="Block">
+
+        </div>
+    )
+}
+
+export default Block
