--- frontend/src/styles/Block.css
@@ -1,5 +1,0 @@
-.Block {
-    background-color: #FFF;
-    border-radius: 45px;
-    margin-bottom: 24px;
-}
