+++ frontend/src/styles/clientAdmin/Block.css
@@ -1,0 +1,5 @@
+.Block {
+    background-color: #FFF;
+    border-radius: 45px;
+    margin-bottom: 24px;
+}
