+++ frontend/src/styles/clientAdmin/Main.css
@@ -1,0 +1,66 @@
+.Main {
+    padding-top: 48px;
+}
+
+.Main_title {
+    display: flex;
+    justify-content: space-between;
+    margin-bottom: 32px;
+}
+
+.Main_titleText {
+    font-family: 'SF-Pro';
+    font-size: 48px;
+    font-weight: 500;
+}
+
+.Main_titleDate {
+    width: 352px;
+    height: 64px;
+    background-color: #EEEEF1;
+    border-radius: 45px;
+    display: flex;
+    justify-content: space-around;
+    align-items: center;
+}
+
+.Main_textDate {
+    font-family: 'SF-Pro';
+}
+
+.Main_calendar {
+    width: 20px;
+    height: 19px;
+}
+
+.Main_arrow {
+    width: 18px;
+    height: 10px;
+}
+
+.Main_Content {
+    display: flex;
+}
+
+.Main_left {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+}
+
+.GraphJobsComp {
+    flex: 2;
+}
+
+.JobComp {
+    flex: 2;
+}
+
+.BlockComp {
+    flex: 1;
+}
+
+.UsersComp {
+    flex: 3;
+    width: 1000px;
+}
